投稿周りリファクタリング
