+++ front/src/app/auth/callback/page.tsx
@@ -22,7 +22,7 @@ function AuthCallbackPageContent() {
     };
 
     setTokenAndClose();
-  }, []);
+  }, [searchParams]);
   return <></>;
 }
