+++ front/src/app/users/[id]/page.tsx
@@ -1,8 +1,9 @@
-export default function User() {
+export default function User({ params: { id } }: { params: { id: string } }) {
   return (
     <article>
       <section>
         <h1>ユーザーページ</h1>
+        <p>ユーザーID: {id}</p>
       </section>
     </article>
   );
