+++ front/src/features/posts/index.ts
@@ -1,5 +1,7 @@
-import DetailModal from "./detailModal";
-import NewLetter from "./newLetter";
-import ReplyLetter from "./reply";
+import { DetailModal } from "./detailModal";
+import { Letter } from "./letter";
+import { LetterDetail } from "./letterDetail";
+import { Letters } from "./letters";
+import PostForm from "./postForm";
 
-export { NewLetter, DetailModal, ReplyLetter };
+export { DetailModal, Letter, Letters, LetterDetail, PostForm };
