+++ front/src/features/posts/letterDetail/index.tsx
@@ -1,0 +1,49 @@
+import { memo, useEffect } from "react";
+import useSWR from "swr";
+
+import { axiosClient } from "@/lib";
+
+const fetcher = (url: string) =>
+  axiosClient()
+    .get(url)
+    .then((res) => res.data);
+
+export const LetterDetail = memo(function LetterDetail({
+  page,
+  uuid,
+  setPageCount,
+}: {
+  page: number;
+  uuid: string;
+  setPageCount?: (value: number) => void;
+}) {
+  const { data } = useSWR(`/posts/${uuid}?page=${page}`, fetcher);
+
+  useEffect(() => {
+    if (data && setPageCount) {
+      setPageCount(data.all_count);
+    }
+  }, []);
+
+  if (data === undefined) {
+    return <p>ちょっとまってね</p>;
+  }
+
+  if (data.letter == null) return;
+
+  return (
+    <>
+      {data === undefined ? (
+        <p>ちょっとまってね</p>
+      ) : (
+        <section className="m-auto my-2 bg-white p-4 px-8">
+          <p className="border-b border-sky-200 text-end">{data.letter.name} より</p>
+          <p className="border-b border-sky-200 pt-1 text-end text-sm text-gray-400">
+            {data.letter.created_at}
+          </p>
+          <p className="lined-textarea whitespace-pre-wrap leading-7">{data.letter.sentences}</p>
+        </section>
+      )}
+    </>
+  );
+});
