+++ front/src/hooks/useDimensions/index.tsx
@@ -10,7 +10,7 @@ export const useDimensions = (ref: RefObject<HTMLElement>) => {
     if (!ref.current) return;
     dimensions.current.width = ref.current.offsetWidth;
     dimensions.current.height = ref.current.offsetHeight;
-  }, []);
+  }, [ref]);
 
   return dimensions.current;
 };
